+++ CHANGELOG.rst
@@ -58,3 +58,4 @@ Changelog
 ===
 
 * Fix variable type bugs
+* Add command line options to control testvars gathering