+++ CHANGELOG.rst
@@ -59,5 +59,6 @@ Changelog
 
 * Fix variable type bugs
 * Add command line options to control testvars gathering
-* Include extra vars (again)
+* Include extra vars (again) as TESTVARS_EXTRA_VARS
 * Provide environment variables as fixtures
+* Start unit tests